[cya] klassBoard 클래스 게시글 목록 정렬 조회 수정 완료 commit
	- 좋아요 순으로 정렬 추가
	- klassBoardUpdate 쪽 오류메세지 부분 filedelete 위치 이동 수정 완료

+++ lumodiem/src/main/java/com/lumodiem/board/hostboard/controller/KlassBoardUpdateEndServlet.java
@@ -146,13 +146,15 @@ public class KlassBoardUpdateEndServlet extends HttpServlet {
 			if(result > 0) {
 				obj.put("res_code", "200");
 				obj.put("res_msg", "정상적으로 게시글이 등록되었습니다.");
+				if(atc != null) {
+					String deletePath = a.getAttachPath();
+					File deleteFile = new File(deletePath);
+					if(deleteFile.exists()) {
+						deleteFile.delete();
+				}
 			}else {
 				obj.put("res_code", "500");
 				obj.put("res_msg", "게시글 등록중 오류가 발생하였습니다.");
-				String deletePath = a.getAttachPath();
-				File deleteFile = new File(deletePath);
-				if(deleteFile.exists()) {
-					deleteFile.delete();
 				}
 				
 			}
